+++ src/middleware.ts
@@ -1,20 +1,22 @@
 import { Middleware, Store } from '@reduxjs/toolkit'
 import CableCar, { CableCarOptions } from './cableCar'
 
-type CableCarMiddleware = Middleware<{}, any> & {
-    connect: (
-        store: Store,
-        channel: string,
-        options: CableCarOptions
-    ) => CableCar
+export type CableCarApi = {
+    createMiddleware: () => Middleware
+    init: (store: Store, channel: string, options: CableCarOptions) => void
+    destroy: () => void
+    perform: (method: string, action: any) => void
+    send: (action: any) => void
 }
 
-const cars: CableCar[] = []
+export function createCableCar(): CableCarApi {
+    let car: null | CableCar = null
 
-const middleware: CableCarMiddleware = (store) => (next) => (action) => {
-    let passActionAlong = true
-    Object.values(cars).forEach((car) => {
-        if (car.permitsAction(action)) {
+    let middleware: Middleware = (store) => (next) => (action) => {
+        let passActionAlong = true
+
+        // runs through logic if action is permitted
+        if (car?.permitsAction(action)) {
             let meta = action.meta || {}
             if (car.active) {
                 passActionAlong = meta.isOptimistic
@@ -25,25 +27,27 @@ const middleware: CableCarMiddleware = (store) => (next) => (action) => {
                     'CableCar: Dropped action.' +
                         (passActionAlong
                             ? ' Action passed thru middleware (optimistic).'
-                            : '')
+                            : ''),
+                    action
                 )
             }
         }
-    })
 
-    return passActionAlong ? next(action) : store.getState()
+        return passActionAlong ? next(action) : store.getState()
+    }
+
+    return {
+        createMiddleware: () => middleware,
+        // CableCar Object api
+        init: (store, channel, options) => {
+            car = new CableCar(store, channel, options)
+        },
+        destroy: () => {
+            car?.destroy()
+            car = null
+        },
+        perform: (method: string, payload: any) =>
+            car?.perform(method, payload),
+        send: (action: any) => car?.send(action),
+    }
 }
-
-middleware.connect = (
-    store: Store,
-    channel: string,
-    options: CableCarOptions
-) => {
-    const car = new CableCar(store, channel, options, () => {
-        cars.splice(cars.indexOf(car, 1))
-    })
-    cars.push(car)
-    return car
-}
-
-export default middleware